[REFACTOR] Remove init adapter in bindingadapter
bindingAdapter에서 했던 recyclerview adapter 연결 fragment에서 연결 하도록 변경
