+++ .idea/misc.xml
@@ -14,6 +14,7 @@
         <entry key="..\:/Users/wodrk/Desktop/Android_ WorkSpace/Sopt-Seminar/app/src/main/res/layout/detail_fragment.xml" value="0.1" />
         <entry key="..\:/Users/wodrk/Desktop/Android_ WorkSpace/Sopt-Seminar/app/src/main/res/layout/follower_fragment.xml" value="0.30757282692521576" />
         <entry key="..\:/Users/wodrk/Desktop/Android_ WorkSpace/Sopt-Seminar/app/src/main/res/layout/follower_frame.xml" value="0.4731800766283525" />
+        <entry key="..\:/Users/wodrk/Desktop/Android_ WorkSpace/Sopt-Seminar/app/src/main/res/layout/follower_list_fragment.xml" value="0.32916666666666666" />
         <entry key="..\:/Users/wodrk/Desktop/Android_ WorkSpace/Sopt-Seminar/app/src/main/res/layout/fragment_camera.xml" value="0.32916666666666666" />
         <entry key="..\:/Users/wodrk/Desktop/Android_ WorkSpace/Sopt-Seminar/app/src/main/res/layout/fragment_follower.xml" value="0.24814814814814815" />
         <entry key="..\:/Users/wodrk/Desktop/Android_ WorkSpace/Sopt-Seminar/app/src/main/res/layout/fragment_following.xml" value="0.3151041666666667" />
