+++ app/src/main/java/com/example/sopt_seminar/ui/RepoFragment.kt
@@ -3,6 +3,8 @@ package com.example.sopt_seminar.ui
 import android.os.Bundle
 import android.view.View
 import androidx.navigation.fragment.findNavController
+import androidx.recyclerview.widget.ItemTouchHelper
+import androidx.recyclerview.widget.RecyclerView
 import com.example.sopt_seminar.R
 import com.example.sopt_seminar.databinding.RepoFragmentBinding
 import com.example.sopt_seminar.domain.model.Repo
@@ -11,33 +13,52 @@ import com.example.sopt_seminar.util.BaseFragment
 import java.util.*
 
 class RepoFragment : BaseFragment<RepoFragmentBinding>(R.layout.repo_fragment) {
+    private val adapter = RepoAdapter { name, description ->
+        val action = MainFragmentDirections.actionMainFragmentToDetailFragment(
+            followerName = name,
+            followerDes = description
+        )
+        findNavController().navigate(action)
+    }
+
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        val adapter = RepoAdapter { name, description ->
-            val action = MainFragmentDirections.actionMainFragmentToDetailFragment(
-                followerName = name,
-                followerDes = description
-            )
-            findNavController().navigate(action)
-        }
-
-        binding.apply {
-            main = this@RepoFragment
-            repoRecyclerView.adapter = adapter
-        }
+        initAdapter()
     }
 
-    val removeItem = fun(position: Int) { testList.removeAt(position) }
+    private fun initAdapter() {
+        val simpleCallback =
+            object : ItemTouchHelper.SimpleCallback(
+                ItemTouchHelper.DOWN or ItemTouchHelper.UP or ItemTouchHelper.START or ItemTouchHelper.END,
+                ItemTouchHelper.LEFT
+            ) {
+                override fun onMove(
+                    recyclerView: RecyclerView,
+                    viewHolder: RecyclerView.ViewHolder,
+                    target: RecyclerView.ViewHolder
+                ): Boolean {
+                    val startPosition = viewHolder.adapterPosition
+                    val endPosition = target.adapterPosition
+                    adapter.moveItem(startPosition, endPosition) {
+                        Collections.swap(testList, startPosition, endPosition)
+                    }
+                    return true
+                }
 
-    val moveItem = fun(fromPosition: Int, toPosition: Int) {
-        Collections.swap(
-            testList,
-            fromPosition,
-            toPosition
-        )
+                override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
+                    when (direction) {
+                        ItemTouchHelper.LEFT -> {
+                            adapter.removeItem(viewHolder.adapterPosition) {
+                                testList.removeAt(viewHolder.adapterPosition)
+                            }
+                        }
+                    }
+                }
+            }
+        binding.repoRecyclerView.adapter = adapter
+        ItemTouchHelper(simpleCallback).attachToRecyclerView(binding.repoRecyclerView)
+        adapter.submitList(testList.toList())
     }
-
-    fun getList(): List<Repo> = testList
 
     companion object {
         private val testList = mutableListOf(
