+++ app/src/main/res/layout/repo_fragment.xml
@@ -2,22 +2,12 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
-    <data>
-
-        <variable
-            name="main"
-            type="com.example.sopt_seminar.ui.RepoFragment" />
-    </data>
-
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/repo_recycler_view"
-            bindList="@{main.getList()}"
-            moveItem="@{main.moveItem}"
-            removeItem="@{main.removeItem}"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
